+++ src/constants.ts
@@ -1,4 +1,4 @@
 import { join } from "path";
 import { homedir } from "os";
 
-export const VAULT_DIR = join(homedir(), ".context-vault");
+export const VAULT_DIR = join(homedir(), ".contexter-vault");